requirements.txt
streamlit
supabase>=2.4.0
groq
python-dotenv
pillow
opencv-python-headless
numpy
pandas